+++ myerpGWT/src/main/java/org/adempiere/web/client/desktop/IDesktop.java
@@ -1,15 +1,15 @@
 package org.adempiere.web.client.desktop;
 
 public interface IDesktop {
-	void openWindow(String name, long iWindowId);
+	void openWindow(String name, Integer iWindowId);
 
-	void openReport(String name, long iProcessId);
+	void openReport(String name, Integer iProcessId);
 
-	void openWorkflow(String name, long iWorkflowId);
+	void openWorkflow(String name, Integer iWorkflowId);
 
-	void openTask(String name, long iTaskId);
+	void openTask(String name, Integer iTaskId);
 
-	void openPorcess(String name, long iProcessId);
+	void openPorcess(String name, Integer iProcessId);
 
-	void openForm(String name, long iFormId);
+	void openForm(String name, Integer iFormId);
 }
